v2

+++ work1.py
@@ -1,6 +1,6 @@
-a = int(input())
-b = int(input())
-c = int(input())
+a = float(input())
+b = float(input())
+c = float(input())
 
 d = (b**2) - (4*a*c)
 x1 = (-b + (d**0.5))/(2*a)
